requirements.txt
discord.py
fastapi
uvicorn
requests
aiofiles
python-dotenv
yt-dlp
PyNaCl